Adding currentTheme into Theme Definition File
Currently I am using Nuxt with TypeScript with Vuetify support in a brand new project.

When I access the object `this.$vuetify.theme`, it displays me a `currentTheme` object. This object has type `Partial<VuetifyThemeVariant>` and I realized it was not in this file.

I don't know if this is intentional but I am proposing the change anyway.

Hope it helps.